+++ contracts/reflect/examples/schema.rs
@@ -2,8 +2,9 @@ use std::env::current_dir;
 use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
+use cosmwasm_std::HandleResponse;
 
-use reflect::msg::{HandleMsg, InitMsg, OwnerResponse, QueryMsg};
+use reflect::msg::{CustomMsg, HandleMsg, InitMsg, OwnerResponse, QueryMsg};
 use reflect::state::State;
 
 fn main() {
@@ -12,8 +13,10 @@ fn main() {
     create_dir_all(&out_dir).unwrap();
     remove_schemas(&out_dir).unwrap();
 
+    export_schema(&schema_for!(CustomMsg), &out_dir);
     export_schema(&schema_for!(InitMsg), &out_dir);
     export_schema(&schema_for!(HandleMsg), &out_dir);
+    export_schema(&schema_for!(HandleResponse<CustomMsg>), &out_dir);
     export_schema(&schema_for!(QueryMsg), &out_dir);
     export_schema(&schema_for!(State), &out_dir);
     export_schema(&schema_for!(OwnerResponse), &out_dir);