+++ contracts/reflect/tests/integration.rs
@@ -5,7 +5,7 @@ use cosmwasm_std::{
 
 use cosmwasm_vm::testing::{handle, init, mock_instance, query, HandleResult};
 
-use reflect::msg::{CustomNativeMsg, HandleMsg, InitMsg, OwnerResponse, QueryMsg};
+use reflect::msg::{CustomMsg, HandleMsg, InitMsg, OwnerResponse, QueryMsg};
 
 /**
 This integration test tries to run and call the generated wasm.
@@ -57,7 +57,7 @@ fn proper_initialization() {
     let env = mock_env(&deps.api, "creator", &coins(1000, "earth"));
 
     // we can just call .unwrap() to assert this was a success
-    let res: InitResponse<CustomNativeMsg> = init(&mut deps, env, msg).unwrap();
+    let res: InitResponse<CustomMsg> = init(&mut deps, env, msg).unwrap();
     assert_eq!(0, res.messages.len());
 
     // it worked, let's query the state
@@ -72,7 +72,7 @@ fn reflect() {
 
     let msg = InitMsg {};
     let env = mock_env(&deps.api, "creator", &coins(2, "token"));
-    let _res: InitResponse<CustomNativeMsg> = init(&mut deps, env, msg).unwrap();
+    let _res: InitResponse<CustomMsg> = init(&mut deps, env, msg).unwrap();
 
     let env = mock_env(&deps.api, "creator", &[]);
     let payload = vec![
@@ -83,13 +83,13 @@ fn reflect() {
         }
         .into(),
         // make sure we can pass through custom native messages
-        CustomNativeMsg::Custom(Binary(b"{\"foo\":123}".to_vec())).into(),
-        CustomNativeMsg::Debug("Hi, Dad!".to_string()).into(),
+        CustomMsg::Raw(Binary(b"{\"foo\":123}".to_vec())).into(),
+        CustomMsg::Debug("Hi, Dad!".to_string()).into(),
     ];
     let msg = HandleMsg::ReflectMsg {
         msgs: payload.clone(),
     };
-    let res: HandleResponse<CustomNativeMsg> = handle(&mut deps, env, msg).unwrap();
+    let res: HandleResponse<CustomMsg> = handle(&mut deps, env, msg).unwrap();
 
     // should return payload
     assert_eq!(payload, res.messages);
@@ -101,7 +101,7 @@ fn reflect_requires_owner() {
 
     let msg = InitMsg {};
     let env = mock_env(&deps.api, "creator", &coins(2, "token"));
-    let _res: InitResponse<CustomNativeMsg> = init(&mut deps, env, msg).unwrap();
+    let _res: InitResponse<CustomMsg> = init(&mut deps, env, msg).unwrap();
 
     // signer is not owner
     let env = mock_env(&deps.api, "someone", &[]);
@@ -115,7 +115,7 @@ fn reflect_requires_owner() {
         msgs: payload.clone(),
     };
 
-    let res: HandleResult<CustomNativeMsg> = handle(&mut deps, env, msg);
+    let res: HandleResult<CustomMsg> = handle(&mut deps, env, msg);
     match res {
         Err(ApiError::Unauthorized {}) => {}
         _ => panic!("Must return unauthorized error"),
@@ -128,14 +128,14 @@ fn transfer() {
 
     let msg = InitMsg {};
     let env = mock_env(&deps.api, "creator", &coins(2, "token"));
-    let _res: InitResponse<CustomNativeMsg> = init(&mut deps, env, msg).unwrap();
+    let _res: InitResponse<CustomMsg> = init(&mut deps, env, msg).unwrap();
 
     let env = mock_env(&deps.api, "creator", &[]);
     let new_owner = HumanAddr::from("friend");
     let msg = HandleMsg::ChangeOwner {
         owner: new_owner.clone(),
     };
-    let res: HandleResponse<CustomNativeMsg> = handle(&mut deps, env, msg).unwrap();
+    let res: HandleResponse<CustomMsg> = handle(&mut deps, env, msg).unwrap();
 
     // should change state
     assert_eq!(0, res.messages.len());
@@ -150,7 +150,7 @@ fn transfer_requires_owner() {
 
     let msg = InitMsg {};
     let env = mock_env(&deps.api, "creator", &coins(2, "token"));
-    let _res: InitResponse<CustomNativeMsg> = init(&mut deps, env, msg).unwrap();
+    let _res: InitResponse<CustomMsg> = init(&mut deps, env, msg).unwrap();
 
     let env = mock_env(&deps.api, "random", &[]);
     let new_owner = HumanAddr::from("friend");